+++ app/src/main/java/org/breezyweather/main/utils/RefreshErrorType.kt
@@ -117,6 +117,9 @@ enum class RefreshErrorType(
     SECONDARY_WEATHER_FAILED(
         shortMessage = R.string.weather_message_secondary_data_refresh_failed
     ),
+    INVALID_INCOMPLETE_DATA(
+        shortMessage = R.string.message_invalid_incomplete_data
+    ),
     WEATHER_REQ_FAILED(
         shortMessage = R.string.weather_message_data_refresh_failed
     );