+++ app/src/main/java/org/breezyweather/sources/brightsky/BrightSkyResultConverter.kt
@@ -29,7 +29,7 @@ import breezyweather.domain.weather.model.Wind
 import breezyweather.domain.weather.wrappers.HourlyWrapper
 import breezyweather.domain.weather.wrappers.SecondaryWeatherWrapper
 import breezyweather.domain.weather.wrappers.WeatherWrapper
-import org.breezyweather.common.exceptions.WeatherException
+import org.breezyweather.common.exceptions.InvalidOrIncompleteDataException
 import org.breezyweather.common.extensions.getFormattedDate
 import org.breezyweather.common.extensions.toDateNoHour
 import org.breezyweather.sources.brightsky.json.BrightSkyAlert
@@ -54,7 +54,7 @@ fun convert(
 ): WeatherWrapper {
     // If the API doesn’t return weather, consider data as garbage and keep cached data
     if (weatherResult.weather.isNullOrEmpty()) {
-        throw WeatherException()
+        throw InvalidOrIncompleteDataException()
     }
 
     return WeatherWrapper(