+++ app/src/src_nofdroid/org/breezyweather/sources/ims/ImsResultConverter.kt
@@ -29,7 +29,7 @@ import breezyweather.domain.weather.model.Wind
 import breezyweather.domain.weather.wrappers.HourlyWrapper
 import breezyweather.domain.weather.wrappers.SecondaryWeatherWrapper
 import breezyweather.domain.weather.wrappers.WeatherWrapper
-import org.breezyweather.common.exceptions.WeatherException
+import org.breezyweather.common.exceptions.InvalidOrIncompleteDataException
 import org.breezyweather.common.extensions.toCalendarWithTimeZone
 import org.breezyweather.common.extensions.toDateNoHour
 import org.breezyweather.sources.ims.json.ImsLocation
@@ -56,7 +56,7 @@ fun convert(
 ): WeatherWrapper {
     // If the API doesn’t return data, consider data as garbage and keep cached data
     if (weatherResult?.data?.forecastData.isNullOrEmpty()) {
-        throw WeatherException()
+        throw InvalidOrIncompleteDataException()
     }
 
     return WeatherWrapper(